# docs/api-reference/openapi.json
{
  "info": {
    "description": "Generated from the live Flask url_map \u2014 see potato/server_utils/openapi_spec.py.\n\nMost endpoints are session-authenticated: POST /auth with `username` and `password` as form data, then reuse the session cookie. `x-potato-auth` records the decorators guarding an operation, and `x-potato-requires-config` marks operations that only exist when the named config is enabled.\n\nResponse schemas are intentionally not modelled; this spec is an authoritative index of what exists, not a payload contract.",
    "title": "Potato HTTP API",
    "version": "2.8.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/arena": {
      "get": {
        "operationId": "get_arena_arena_page",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "tags": [
          "arena"
        ],
        "x-potato-auth": [
          "admin_required"
        ],
        "x-potato-requires-config": "arena.enabled: true"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}